// react-module/src/openapi/models/machine-calendar-task-extended-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface MachineCalendarTaskExtendedDto
 */
export interface MachineCalendarTaskExtendedDto {
    /**
     * 
     * @type {number}
     * @memberof MachineCalendarTaskExtendedDto
     */
    id?: any;
    /**
     * 
     * @type {Date}
     * @memberof MachineCalendarTaskExtendedDto
     */
    startDate?: any;
    /**
     * 
     * @type {Date}
     * @memberof MachineCalendarTaskExtendedDto
     */
    startDatePlanned?: any;
    /**
     * 
     * @type {Date}
     * @memberof MachineCalendarTaskExtendedDto
     */
    endDate?: any;
    /**
     * 
     * @type {Date}
     * @memberof MachineCalendarTaskExtendedDto
     */
    endDatePlanned?: any;
    /**
     * 
     * @type {Array&lt;AtomicOperationExtendedDto&gt;}
     * @memberof MachineCalendarTaskExtendedDto
     */
    operationList?: any;
    /**
     * 
     * @type {OperationTypeExtendedDto}
     * @memberof MachineCalendarTaskExtendedDto
     */
    operationType?: any;
    /**
     * 
     * @type {MachineDto}
     * @memberof MachineCalendarTaskExtendedDto
     */
    machine?: any;
    /**
     * 
     * @type {Array&lt;UserWorkTime&gt;}
     * @memberof MachineCalendarTaskExtendedDto
     */
    userWorkTimes?: any;
    /**
     * 
     * @type {string}
     * @memberof MachineCalendarTaskExtendedDto
     */
    planningStatus?: MachineCalendarTaskExtendedDtoPlanningStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof MachineCalendarTaskExtendedDto
     */
    productionStatus?: MachineCalendarTaskExtendedDtoProductionStatusEnum;
    /**
     * 
     * @type {ForSavingEstimatedTime}
     * @memberof MachineCalendarTaskExtendedDto
     */
    estimatedTime?: any;
    /**
     * 
     * @type {string}
     * @memberof MachineCalendarTaskExtendedDto
     */
    info?: any;
    /**
     * 
     * @type {SurfaceWithPlacedEntities}
     * @memberof MachineCalendarTaskExtendedDto
     */
    spe?: any;
}

/**
    * @export
    * @enum {string}
    */
export enum MachineCalendarTaskExtendedDtoPlanningStatusEnum {
    INIT = 'INIT',
    PLANNED = 'PLANNED',
    CONFIRMED = 'CONFIRMED',
    FINISHED = 'FINISHED',
    DELETED = 'DELETED'
}
/**
    * @export
    * @enum {string}
    */
export enum MachineCalendarTaskExtendedDtoProductionStatusEnum {
    INIT = 'INIT',
    INPROGRESS = 'IN_PROGRESS',
    POSTPONED = 'POSTPONED',
    FINISHED = 'FINISHED',
    DELETED = 'DELETED',
    FAULT = 'FAULT'
}